src/VumbaSoft.AdventureWorks.Controllers/Purchasing/ShipMethods/ShipMethodsController.cs
using Microsoft.AspNetCore.Mvc;
using VumbaSoft.AdventureWorks.Objects;
using VumbaSoft.AdventureWorks.Services;
using VumbaSoft.AdventureWorks.Validators;
using System;

namespace VumbaSoft.AdventureWorks.Controllers.Purchasing
{
    [Area("Purchasing")]
    public class ShipMethodsController : ValidatedController<IShipMethodValidator, IShipMethodService>
    {
        public ShipMethodsController(IShipMethodValidator validator, IShipMethodService service)
            : base(validator, service)
        {
        }

        [HttpGet]
        public ViewResult Index()
        {
            return View(Service.GetViews());
        }

        [HttpGet]
        public ViewResult Create()
        {
            return View();
        }

        [HttpPost]
        public ActionResult Create(ShipMethodView method)
        {
            if (!Validator.CanCreate(method))
                return View(method);

            Service.Create(method);

            return RedirectToAction("Index");
        }

        [HttpGet]
        public ActionResult Details(Int32 id)
        {
            return NotEmptyView(Service.Get<ShipMethodView>(id));
        }

        [HttpGet]
        public ActionResult Edit(Int32 id)
        {
            return NotEmptyView(Service.Get<ShipMethodView>(id));
        }

        [HttpPost]
        public ActionResult Edit(ShipMethodView method)
        {
            if (!Validator.CanEdit(method))
                return View(method);

            Service.Edit(method);

            return RedirectToAction("Index");
        }

        [HttpGet]
        public ActionResult Delete(Int32 id)
        {
            return NotEmptyView(Service.Get<ShipMethodView>(id));
        }

        [HttpPost]
        [ActionName("Delete")]
        public RedirectToActionResult DeleteConfirmed(Int32 id)
        {
            Service.Delete(id);

            return RedirectToAction("Index");
        }
    }
}
